+++ apps/enmeshed/lib/core/modals/modals.dart
@@ -1,5 +1,6 @@
 export 'create_attribute.dart';
 export 'delete_attribute.dart';
+export 'delete_local_data.dart';
 export 'enter_password/enter_password.dart';
 export 'restore_identity.dart';
 export 'succeed_attribute.dart';
