+++ apps/enmeshed/lib/core/widgets/widgets.dart
@@ -8,6 +8,7 @@ export 'file_icon.dart';
 export 'file_item.dart';
 export 'highlight_text.dart';
 export 'instructions_screen.dart';
+export 'local_data_deletion/local_data_deletion.dart';
 export 'message_dvo_renderer.dart';
 export 'messages_container.dart';
 export 'modal_loading_overlay.dart';
