+++ apps/enmeshed/lib/profiles/profile/modals/modals.dart
@@ -1,2 +1,2 @@
-export 'delete_local_data_modal.dart';
+export 'edit_profile/edit_profile.dart';
 export 'edit_profile/edit_profile_modal.dart';
