+++ apps/enmeshed/lib/profiles/profile/profiles_screen.dart
@@ -358,51 +358,9 @@ class _ProfilesInDeletion extends StatelessWidget {
         ),
         if (accountsInDeletion.any((e) => e.deletionDate != null)) ...[
           Gaps.h16,
-          _DeleteDataNowCard(reloadAccounts: reloadAccounts, accountsInDeletion: accountsInDeletion.where((e) => e.deletionDate != null).toList()),
+          DeleteDataNowCard(reloadAccounts: reloadAccounts, accountsInDeletion: accountsInDeletion.where((e) => e.deletionDate != null).toList()),
         ]
       ],
-    );
-  }
-}
-
-class _DeleteDataNowCard extends StatelessWidget {
-  final List<LocalAccountDTO> accountsInDeletion;
-  final VoidCallback reloadAccounts;
-
-  const _DeleteDataNowCard({required this.accountsInDeletion, required this.reloadAccounts});
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 16),
-      child: Card.filled(
-        child: Column(
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(16),
-              child: Row(
-                children: [
-                  Icon(Icons.error, color: Theme.of(context).colorScheme.error),
-                  Gaps.w8,
-                  Expanded(child: Text(context.l10n.profile_localDeletion_title, style: Theme.of(context).textTheme.bodyMedium)),
-                ],
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 16),
-              child: Text(context.l10n.profile_localDeletion_card_description, style: Theme.of(context).textTheme.bodySmall),
-            ),
-            Padding(
-              padding: const EdgeInsets.symmetric(vertical: 24),
-              child: OutlinedButton.icon(
-                onPressed: () => showDeleteLocalDataModal(context: context, accountsInDeletion: accountsInDeletion, onDeleted: reloadAccounts),
-                label: Text(context.l10n.profile_localDeletion_card_button),
-                icon: Icon(Icons.delete_forever_outlined),
-              ),
-            ),
-          ],
-        ),
-      ),
     );
   }
 }
